+++ src/http/mod.rs
@@ -5,21 +5,17 @@ pub use http::uri::{Authority, PathAndQuery, Uri};
 
 #[doc(inline)]
 pub use body::{Body, util::BodyExt};
-pub use client::Client;
+pub use crate::sys::http::client::Client;
 pub use error::{Error, ErrorCode, Result};
-pub use fields::{HeaderMap, HeaderName, HeaderValue};
-pub use method::Method;
+pub use crate::sys::http::fields::{HeaderMap, HeaderName, HeaderValue};
+pub use crate::sys::http::method::Method;
 pub use request::Request;
 pub use response::Response;
-pub use scheme::{InvalidUri, Scheme};
+pub use crate::sys::http::scheme::{InvalidUri, Scheme};
 
 pub mod body;
 
-mod client;
 pub mod error;
-mod fields;
-mod method;
 pub mod request;
 pub mod response;
-mod scheme;
 pub mod server;
