+++ src/runtime/mod.rs
@@ -10,27 +10,4 @@
 #![deny(missing_debug_implementations, nonstandard_style)]
 #![warn(missing_docs, unreachable_pub)]
 
-mod block_on;
-mod reactor;
-
-pub use ::async_task::Task;
-pub use block_on::block_on;
-pub use reactor::{AsyncPollable, Reactor, WaitFor};
-use std::cell::RefCell;
-
-// There are no threads in WASI 0.2, so this is just a safe way to thread a single reactor to all
-// use sites in the background.
-std::thread_local! {
-pub(crate) static REACTOR: RefCell<Option<Reactor>> = const { RefCell::new(None) };
-}
-
-/// Spawn a `Future` as a `Task` on the current `Reactor`.
-///
-/// Panics if called from outside `block_on`.
-pub fn spawn<F, T>(fut: F) -> Task<T>
-where
-    F: std::future::Future<Output = T> + 'static,
-    T: 'static,
-{
-    Reactor::current().spawn(fut)
-}
+pub use crate::sys::runtime::*;
