+++ src/sys/p2/http/client.rs
@@ -1,6 +1,6 @@
-use super::{Body, Error, Request, Response};
-use crate::http::request::try_into_outgoing;
-use crate::http::response::try_from_incoming;
+use crate::http::{Body, Error, Request, Response};
+use super::request::try_into_outgoing;
+use super::response::try_from_incoming;
 use crate::io::AsyncPollable;
 use crate::time::Duration;
 use wasip2::http::types::RequestOptions as WasiRequestOptions;
