+++ src/sys/p2/http/mod.rs
@@ -1,0 +1,10 @@
+pub mod body;
+pub(crate) mod client;
+pub(crate) mod fields;
+pub(crate) mod method;
+pub mod request;
+pub mod response;
+pub(crate) mod scheme;
+pub mod server;
+
+pub use wasip2::http::types::{ErrorCode, HeaderError};
