+++ src/sys/p2/net/mod.rs
@@ -1,0 +1,27 @@
+use std::io::{self, ErrorKind};
+use wasip2::sockets::network::ErrorCode;
+
+mod tcp_listener;
+mod tcp_stream;
+
+pub use tcp_listener::*;
+pub use tcp_stream::*;
+
+pub(crate) fn to_io_err(err: ErrorCode) -> io::Error {
+    match err {
+        ErrorCode::Unknown => ErrorKind::Other.into(),
+        ErrorCode::AccessDenied => ErrorKind::PermissionDenied.into(),
+        ErrorCode::NotSupported => ErrorKind::Unsupported.into(),
+        ErrorCode::InvalidArgument => ErrorKind::InvalidInput.into(),
+        ErrorCode::OutOfMemory => ErrorKind::OutOfMemory.into(),
+        ErrorCode::Timeout => ErrorKind::TimedOut.into(),
+        ErrorCode::WouldBlock => ErrorKind::WouldBlock.into(),
+        ErrorCode::InvalidState => ErrorKind::InvalidData.into(),
+        ErrorCode::AddressInUse => ErrorKind::AddrInUse.into(),
+        ErrorCode::ConnectionRefused => ErrorKind::ConnectionRefused.into(),
+        ErrorCode::ConnectionReset => ErrorKind::ConnectionReset.into(),
+        ErrorCode::ConnectionAborted => ErrorKind::ConnectionAborted.into(),
+        ErrorCode::ConcurrencyConflict => ErrorKind::AlreadyExists.into(),
+        _ => ErrorKind::Other.into(),
+    }
+}
